local insert funzionante dioooo (tolta la strategy cache) e migliorata la random

+++ libs/src/classes/crdt/kk_crdt.h
@@ -30,10 +30,10 @@ class kk_crdt {
 public:
     vector<list<kk_char_ptr>> text;
     string siteid;
-    int boundary;
+    unsigned long boundary;
     strategy _strategy;
     vector<strategy> strategy_cache;
-    int base;
+    unsigned long base;
 
     kk_crdt(string siteid, strategy strategy); //costruttore
     void local_insert(char val, kk_pos pos);
@@ -45,11 +45,11 @@ public:
     vector<kk_identifier_ptr> find_position_after(kk_pos pos);//trova la position della Char immediatamente dopo di quella passata
 
     vector<kk_identifier_ptr> generate_position_between(vector<kk_identifier_ptr> position_1, vector<kk_identifier_ptr> position_2,
-                                                    vector<kk_identifier_ptr> *new_position, int livello); //partendo dalle position di due Char(adiacenti) genera la posizione della Char
+                                                    vector<kk_identifier_ptr> *new_position, unsigned long livello); //partendo dalle position di due Char(adiacenti) genera la posizione della Char
 
-    strategy find_strategy(int level);//trova la strategia migliore per assegnare un identifier alla position della nuova Char
+    strategy find_strategy(unsigned long level);//trova la strategia migliore per assegnare un identifier alla position della nuova Char
 
-    double generate_identifier_between(int min, int max, strategy _strategy); //dati due identifier ne genera uno nuovo da mettere nella position della nuova Char usando la strategia opportuna
+    unsigned long generate_identifier_between(unsigned long min, unsigned long max, strategy _strategy); //dati due identifier ne genera uno nuovo da mettere nella position della nuova Char usando la strategia opportuna
 
 //    void handle_remote_insert(kk_char _char);
 //    kk_pos find_insert_position(kk_char _char);
